\name{BLOSUM62}
\alias{BLOSUM62}
\docType{data}
\title{
  BLOSUM62
}
\description{
  Scoring matrix BLOSUM62
}
\usage{data("BLOSUM62")}
\format{
  A data frame with 25 observations on the following 26 variables.
  \describe{
    \item{\code{X.1}}{a character vector}
    \item{\code{A}}{a numeric vector}
    \item{\code{R}}{a numeric vector}
    \item{\code{N}}{a numeric vector}
    \item{\code{D}}{a numeric vector}
    \item{\code{C}}{a numeric vector}
    \item{\code{Q}}{a numeric vector}
    \item{\code{E}}{a numeric vector}
    \item{\code{G}}{a numeric vector}
    \item{\code{H}}{a numeric vector}
    \item{\code{I}}{a numeric vector}
    \item{\code{L}}{a numeric vector}
    \item{\code{K}}{a numeric vector}
    \item{\code{M}}{a numeric vector}
    \item{\code{F}}{a numeric vector}
    \item{\code{P}}{a numeric vector}
    \item{\code{S}}{a numeric vector}
    \item{\code{T}}{a numeric vector}
    \item{\code{W}}{a numeric vector}
    \item{\code{Y}}{a numeric vector}
    \item{\code{V}}{a numeric vector}
    \item{\code{B}}{a numeric vector}
    \item{\code{J}}{a numeric vector}
    \item{\code{Z}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
    \item{\code{X.}}{a numeric vector}
  }
}
\source{
  Biostrings::data("BLOSUM62")
}
\keyword{datasets}
