% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dict.R
\name{build.dict}
\alias{build.dict}
\title{Build a hashtable with all possible ids with empty score}
\usage{
build.dict(idlist)
}
\arguments{
\item{idlist}{list of possible id resulting from the generate_id function}
}
\value{
a hash table
}
\description{
Build a hashtable with all possible ids with empty score
}
