% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score.R
\name{calculate.score}
\alias{calculate.score}
\title{Calculate the sscore for possible predecessors.}
\usage{
calculate.score(current, predecessoridlist, listofseq, scoremat)
}
\arguments{
\item{current}{id of the current cell}

\item{predecessoridlist}{list of the id of the possible predecessors}

\item{listofseq}{list of sequence strings}

\item{scoremat}{matrix of substitution score}
}
\value{
a hash table of the predecessor - sscore
}
\description{
the function returns a hash table of the pair predecessor - sscore
}
