% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_id.R
\name{generate.id}
\alias{generate.id}
\title{Generate cells id}
\usage{
generate.id(listofseqlength)
}
\arguments{
\item{listofseqlength}{a vector of sequence lengths}
}
\description{
receive a vector of sequence lengths and return the all possible combinations of score id
}
