% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msa.R
\name{getMSA}
\alias{getMSA}
\title{Returns multiple sequence alignment}
\usage{
getMSA(path, seqlist)
}
\arguments{
\item{path}{list of ids representing a path of backward travel}

\item{seqlist}{list of sequences}
}
\description{
Returns multiple sequence alignment
}
