% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goback.R
\name{goback}
\alias{goback}
\title{Find the cell to trace back to}
\usage{
goback(id, gapindices)
}
\arguments{
\item{id}{current cell id}

\item{gapindices}{gap indices}
}
\description{
Find the cell to trace back to
}
