% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtotable.R
\name{hash.astable}
\alias{hash.astable}
\title{display hash table as a dataframe}
\usage{
hash.astable(hashtable)
}
\arguments{
\item{hashtable}{a hash table as a result from hashtab function}
}
\description{
display hash table as a dataframe
}
