% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_score.R
\name{initiate.score}
\alias{initiate.score}
\title{Generate the scores along the axes}
\usage{
initiate.score(dict, listofseqlength)
}
\arguments{
\item{dict}{hashtable of cell id and score}

\item{listofseqlength}{vector of sequence length}
}
\description{
Generate the scores along the axes
}
