% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read.fasta}
\alias{read.fasta}
\title{Read fasta file into table of seqname and sequences}
\usage{
read.fasta(filename)
}
\arguments{
\item{filename}{name of the fasta file}
}
\description{
Read fasta file into table of seqname and sequences
}
