% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_seqlengths.R
\name{return.seqlengths}
\alias{return.seqlengths}
\title{Get sequence length vectors}
\usage{
return.seqlengths(seqtable)
}
\arguments{
\item{seqtable}{a table with column 1 being seqname and 2 being sequence}
}
\description{
Get sequence length vectors
}
