% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_path.R
\name{returnPath}
\alias{returnPath}
\title{Generate trace back path for MSA visualization}
\usage{
returnPath(dict.pred)
}
\arguments{
\item{dict.pred}{a hash table of cellid:predecessor}
}
\value{
path
}
\description{
Generate trace back path for MSA visualization
}
