% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_score.R
\name{s.score}
\alias{s.score}
\title{Calculate the sscore of a column in the MSA table}
\usage{
s.score(scoremat, resvec, gappen)
}
\arguments{
\item{scoremat}{scoring matrix}

\item{resvec}{list of residues in that column}

\item{gappen}{vector of gap penalties}
}
\value{
sscore
}
\description{
Calculate the sscore of a column in the MSA table
}
