% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_score.R
\name{update.score}
\alias{update.score}
\title{update the score of a cell}
\usage{
\method{update}{score}(current, dict, dict.pred, sscoretab)
}
\arguments{
\item{current}{current cell id whose score needs to be filled}

\item{dict}{hashtable containing cellid-score, to retrieve Fscore}

\item{dict.pred}{hashtable containing cellid-predecessor}

\item{sscoretab}{hashtable of predecessor-sscore}
}
\description{
update the score of a cell
}
