% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_command.R
\name{write.command}
\alias{write.command}
\title{Write the command to integrate all functions}
\usage{
write.command()
}
\description{
writes the script after accepting the name of the fasta file containing the sequence
}
